analytics-service/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1,<4.0
django-cors-headers>=4.3,<5.0
dj-database-url>=2.1,<3.0
PyJWT>=2.8,<3.0
